+++ flakytests/restart/21-task-elapsed.t
@@ -39,9 +39,9 @@ run_ok "${TEST_NAME_BASE}-validate" cylc validate "${SUITE_NAME}"
 
 RUND="$(cylc get-global-config --print-run-dir)/${SUITE_NAME}"
 suite_run_ok "${TEST_NAME_BASE}-run" \
-    timeout 120 cylc run "${SUITE_NAME}" --debug --no-detach
+    cylc run "${SUITE_NAME}" --debug --no-detach
 suite_run_ok "${TEST_NAME_BASE}-restart-1" \
-    timeout 120 cylc restart "${SUITE_NAME}" --until=2028 --debug --no-detach
+    cylc restart "${SUITE_NAME}" --until=2028 --debug --no-detach
 sed -n '/LOADING task run times/,+2{s/^.* INFO - //;s/[0-9]\(,\|$\)/%d\1/g;p}' \
     "${RUND}/log/suite/log" >'restart-1.out'
 contains_ok 'restart-1.out' <<'__OUT__'
@@ -50,7 +50,7 @@ LOADING task run times
 + t1: %d,%d,%d,%d,%d
 __OUT__
 suite_run_ok "${TEST_NAME_BASE}-restart-2" \
-    timeout 120 cylc restart "${SUITE_NAME}" --until=2030 --debug --no-detach
+    cylc restart "${SUITE_NAME}" --until=2030 --debug --no-detach
 sed -n '/LOADING task run times/,+2{s/^.* INFO - //;s/[0-9]\(,\|$\)/%d\1/g;p}' \
     "${RUND}/log/suite/log" >'restart-2.out'
 contains_ok 'restart-2.out' <<'__OUT__'
@@ -59,14 +59,14 @@ LOADING task run times
 + t1: %d,%d,%d,%d,%d,%d,%d,%d,%d,%d
 __OUT__
 suite_run_ok "${TEST_NAME_BASE}-restart-3" \
-    timeout 120 cylc restart "${SUITE_NAME}" --until=2031 --hold
+    cylc restart "${SUITE_NAME}" --until=2031 --hold
 # allow the task pool to settle before requesting a dump
 cylc suite-state "${SUITE_NAME}" \
     --task=t1 \
     --point=2031 \
     --status=running \
     --interval=1 \
-    --max-polls=10
+    --max-polls=10 1>'/dev/null' 2>&1
 cylc dump -r "${SUITE_NAME}" >'cylc-dump.out'
 test_dump 'cylc-dump.out'