+++ flakytests/shutdown/02-no-dir.t
@@ -17,7 +17,7 @@
 #-------------------------------------------------------------------------------
 # Test suite can shutdown successfully if its run dir is deleted
 . "$(dirname "$0")/test_header"
-set_test_number 3
+set_test_number 4
 install_suite "${TEST_NAME_BASE}" "${TEST_NAME_BASE}"
 
 run_ok "${TEST_NAME_BASE}-validate" cylc validate "${SUITE_NAME}"
@@ -26,8 +26,8 @@ SYM_SUITE_RUND="${SUITE_RUN_DIR}-sym"
 SYM_SUITE_NAME="${SUITE_NAME}-sym"
 ln -s "$(basename "${SUITE_NAME}")" "${SYM_SUITE_RUND}"
 run_fail "${TEST_NAME_BASE}-run" cylc run "${SYM_SUITE_NAME}" --debug --no-detach
-grep_ok 'CRITICAL - Suite shutting down - unable to open database file' \
-    "${SUITE_RUN_DIR}/log/suite/log".*
+grep_ok 'CRITICAL - Suite shutting down' "${SUITE_RUN_DIR}/log/suite/log".*
+grep_ok 'unable to open database file' "${SUITE_RUN_DIR}/log/suite/log".*
 
 rm -f "${SYM_SUITE_RUND}"
 purge_suite "${SUITE_NAME}"