+++ tests/broadcast/00-simple.t
@@ -18,19 +18,21 @@
 # Test broadcasts
 . "$(dirname "$0")/test_header"
 
-set_test_number 4
+set_test_number 5
 install_suite "${TEST_NAME_BASE}" "${TEST_NAME_BASE}"
 
 run_ok "${TEST_NAME_BASE}-validate" cylc validate "${SUITE_NAME}"
 suite_run_ok "${TEST_NAME_BASE}-run" \
     cylc run --debug --no-detach --reference-test "${SUITE_NAME}"
+sort 'broadcast.log' >'broadcast.log.sorted'
+cmp_ok 'broadcast.ref' 'broadcast.log.sorted'
 
 if ! which sqlite3 > /dev/null; then
     skip 2 "sqlite3 not installed?"
     exit 0
 fi
 
-DB_FILE="$(cylc get-global-config '--print-run-dir')/${SUITE_NAME}/log/db"
+DB_FILE="${SUITE_RUN_DIR}/log/db"
 NAME='select-broadcast-events.out'
 sqlite3 "${DB_FILE}" \
     'SELECT change, point, namespace, key, value FROM broadcast_events