+++ tests/cyclers/40-integer_exclusions_advanced.t
@@ -16,32 +16,15 @@
 # along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #-------------------------------------------------------------------------------
 # Test intercycle dependencies.
-. $(dirname $0)/test_header
+. "$(dirname "$0")/test_header"
+set_test_number 3
+install_suite "${TEST_NAME_BASE}" 'integer_exclusions_advanced'
 #-------------------------------------------------------------------------------
-if [[ -f "$TEST_SOURCE_DIR/$TEST_NAME_BASE-find.out" ]]; then
-    set_test_number 4
-else
-    set_test_number 3
-fi
+run_ok "${TEST_NAME_BASE}-validate" cylc validate "${SUITE_NAME}"
+graph_suite "${SUITE_NAME}" "${SUITE_NAME}.graph.plain"
+cmp_ok "${SUITE_NAME}.graph.plain" 'graph.plain.ref'
+suite_run_ok "${TEST_NAME_BASE}-run" \
+    cylc run --reference-test --debug --no-detach "${SUITE_NAME}"
 #-------------------------------------------------------------------------------
-CHOSEN_SUITE=$(basename $0 | sed "s/^.*-\(.*\)\.t/\1/g")
-install_suite $TEST_NAME_BASE $CHOSEN_SUITE
-#-------------------------------------------------------------------------------
-TEST_NAME=$TEST_NAME_BASE-validate
-run_ok $TEST_NAME cylc validate "$SUITE_NAME"
-#-------------------------------------------------------------------------------
-TEST_NAME=$TEST_NAME_BASE-graph
-graph_suite "$SUITE_NAME" "$SUITE_NAME.graph.plain" 
-cmp_ok "$SUITE_NAME.graph.plain" "$TEST_SOURCE_DIR/$CHOSEN_SUITE/graph.plain.ref"
-#-------------------------------------------------------------------------------
-TEST_NAME=$TEST_NAME_BASE-run
-suite_run_ok $TEST_NAME cylc run --reference-test --debug --no-detach $SUITE_NAME
-#-------------------------------------------------------------------------------
-if [[ -f "$TEST_SOURCE_DIR/$TEST_NAME_BASE-find.out" ]]; then
-    TEST_NAME="$TEST_NAME_BASE-find"
-    SUITE_DIR="$(cylc get-global-config --print-run-dir)/$SUITE_NAME"
-    (cd "$SUITE_DIR"; find log/job work -type f | sort -V) >"$TEST_NAME"
-    cmp_ok "$TEST_NAME" "$TEST_SOURCE_DIR/$TEST_NAME_BASE-find.out"
-fi
-#-------------------------------------------------------------------------------
-purge_suite $SUITE_NAME
+purge_suite "${SUITE_NAME}"
+exit