+++ tests/registration/02-on-the-fly.t
@@ -42,9 +42,9 @@ contains_ok "${TEST_NAME}-run.stdout" <<__ERR__
 REGISTERED ${TESTD} -> ${PWD}
 __ERR__
 
-run_ok "${TEST_NAME}-stop" cylc stop "${TESTD}"
+run_ok "${TEST_NAME}-stop" cylc stop --max-polls=10 --interval=2 "${TESTD}"
 
-purge_suite $TESTD
+purge_suite "${TESTD}"
 #------------------------------------------------------------------------------
 # Test `cylc run` REG for an un-registered suite
 TESTD="cylctb-${CYLC_TEST_TIME_INIT}/${TEST_NAME_BASE}"
@@ -68,9 +68,9 @@ contains_ok "${TEST_NAME}-run.stdout" <<__ERR__
 REGISTERED ${TESTD} -> ${CYLC_RUN_DIR}/${TESTD}
 __ERR__
 
-run_ok "${TEST_NAME}stop-" cylc stop "${TESTD}"
+run_ok "${TEST_NAME}-stop" cylc stop  --max-polls=10 --interval=2 "${TESTD}"
 
-purge_suite $TESTD
+purge_suite "${TESTD}"
 #------------------------------------------------------------------------------
 # Test `cylc run` REG for an un-registered suite
 mkdir -p "${CYLC_RUN_DIR}/${TESTD}"
@@ -94,6 +94,6 @@ contains_ok "${TEST_NAME}-validate.stderr" <<__ERR__
 IllegalItemError: sched
 __ERR__
 
-purge_suite $TESTD
+purge_suite "${TESTD}"
 
 exit