+++ tests/remote/basic/bin/cleanup.sh
@@ -1,9 +1,0 @@
-#!/bin/bash
-# remove installed remote passphrase
-set -e
-cylc check-triggering "$@"
-if [[ $CYLC_TEST_OWNER@$CYLC_TEST_HOST == $USER@localhost || \
-      $CYLC_TEST_OWNER@$CYLC_TEST_HOST == $USER@$(hostname) ]]; then
-    exit
-fi
-echo "Done"