+++ tests/restart/10-pre-initial-2.t
@@ -17,20 +17,20 @@
 #-------------------------------------------------------------------------------
 # Test restarting a suite with pre-initial cycle dependencies and no
 # initial cycle point in suite definition, ref. github #957.
-. $(dirname $0)/test_header
-#-------------------------------------------------------------------------------
+. "$(dirname "$0")/test_header"
 set_test_number 3
+install_suite "${TEST_NAME_BASE}" 'pre-init-2'
 #-------------------------------------------------------------------------------
-install_suite $TEST_NAME_BASE pre-init-2
-export TEST_DIR
+ICP='20100808T00'
+run_ok "${TEST_NAME_BASE}-validate" \
+    cylc validate --initial-cycle-point="${ICP}" "${SUITE_NAME}"
+suite_run_ok "${TEST_NAME_BASE}-run" \
+    cylc run --debug --no-detach \
+    --initial-cycle-point="${ICP}" \
+    --stop-cycle-point="${ICP}" \
+    "${SUITE_NAME}"
+suite_run_ok "${TEST_NAME_BASE}-restart" \
+    cylc restart --debug --no-detach --reference-test "${SUITE_NAME}"
 #-------------------------------------------------------------------------------
-TEST_NAME=$TEST_NAME_BASE-validate
-run_ok $TEST_NAME cylc validate --icp=20100808T00 $SUITE_NAME
-#-------------------------------------------------------------------------------
-TEST_NAME=$TEST_NAME_BASE-run
-suite_run_ok $TEST_NAME cylc run --debug --no-detach --until=20100808T00 $SUITE_NAME 20100808T00
-#-------------------------------------------------------------------------------
-TEST_NAME=$TEST_NAME_BASE-restart
-suite_run_ok $TEST_NAME cylc restart --debug --no-detach --reference-test $SUITE_NAME
-#-------------------------------------------------------------------------------
-purge_suite $SUITE_NAME
+purge_suite "${SUITE_NAME}"
+exit