+++ tests/restart/43-auto-restart-force-override-normal.t
@@ -17,13 +17,14 @@
 #-------------------------------------------------------------------------------
 # Check that "Force Mode" can override a scheduler "Normal Mode" restart.
 . "$(dirname "$0")/test_header"
-export CYLC_TEST_HOST_2=$( \
+CYLC_TEST_HOST_2="$( \
     cylc get-global-config -i '[test battery]remote host with shared fs' \
-    2>'/dev/null')
+    2>'/dev/null')"
 if [[ -z "${CYLC_TEST_HOST_2}" ]]; then
     skip_all '"[test battery]remote host with shared fs": not defined'
 fi
-export CYLC_TEST_HOST_1="$(hostname)"
+export CYLC_TEST_HOST_2
+export CYLC_TEST_HOST_1="${HOSTNAME}"
 
 BASE_GLOBALRC='
 [cylc]
@@ -37,7 +38,8 @@ BASE_GLOBALRC='
 
 TEST_DIR="$HOME/cylc-run/" init_suite "${TEST_NAME_BASE}" <<< '
 [cylc]
-    abort if any task fails = True
+    [[events]]
+        abort if any task fails = True
 [scheduling]
     initial cycle point = 2000
     [[graph]]